+++ tests/E2E/Shop/Checkout/TpayCreditCardCheckoutTest.php
@@ -7,7 +7,8 @@ namespace Tests\CommerceWeavers\SyliusTpayPlugin\E2E\Shop\Checkout;
 use Tests\CommerceWeavers\SyliusTpayPlugin\E2E\E2ETestCase;
 use Tests\CommerceWeavers\SyliusTpayPlugin\E2E\Helper\Account\LoginShopUserTrait;
 use Tests\CommerceWeavers\SyliusTpayPlugin\E2E\Helper\Order\CartTrait;
-use Tests\CommerceWeavers\SyliusTpayPlugin\E2E\Helper\Order\NotificationTrait;
+use Tests\CommerceWeavers\SyliusTpayPlugin\E2E\Helper\Order\PaymentNotificationTrait;
+use Tests\CommerceWeavers\SyliusTpayPlugin\E2E\Helper\Order\TpayNotificationPayloadTrait;
 use Tests\CommerceWeavers\SyliusTpayPlugin\E2E\Helper\Order\TpayTrait;
 
 final class TpayCreditCardCheckoutTest extends E2ETestCase
@@ -15,7 +16,8 @@ final class TpayCreditCardCheckoutTest extends E2ETestCase
     use CartTrait;
     use TpayTrait;
     use LoginShopUserTrait;
-    use NotificationTrait;
+    use PaymentNotificationTrait;
+    use TpayNotificationPayloadTrait;
 
     private const FORM_ID = 'sylius_checkout_complete';
 
@@ -41,8 +43,9 @@ final class TpayCreditCardCheckoutTest extends E2ETestCase
 
         self::assertPageTitleContains('Waiting for payment');
 
-        $this->sendNotificationForLastPayment($this->getCardNotificationPayload());
-        $this->client->executeScript("window.location.reload()");
+        $this->triggerNotificationForLastPayment($this->getCreditCardNotificationPayload());
+
+        $this->client->waitForElementToContain('body', 'Thank you!');
 
         self::assertPageTitleContains('Thank you!');
     }
@@ -53,25 +56,9 @@ final class TpayCreditCardCheckoutTest extends E2ETestCase
         $this->fillCardData(self::FORM_ID, self::CARD_NUMBER, '123', '01', '2029', true);
         $this->placeOrder();
 
-        $this->sendNotificationForLastPayment($this->getCardNotificationPayload());
+        $this->triggerNotificationForLastPayment($this->getCreditCardNotificationPayload());
         $this->client->get('/en_US/account/credit-cards');
 
         self::assertSelectorTextContains('[data-test-grid-table-body]', 'VISA 1111 01-2029');
-    }
-
-    private function getCardNotificationPayload(): array
-    {
-        return [
-            'tr_status' => 'TRUE',
-            'id' => '12345',
-            'tr_id' => 'TR-TEST-123',
-            'tr_amount' => '10.00',
-            'tr_crc' => 'crc',
-            'md5sum' => md5('dummy'),
-            'card_token' => 'tok_abcdef',
-            'card_brand' => 'VISA',
-            'card_tail' => '1111',
-            'token_expiry_date' => '0129',
-        ];
     }
 }